fastapi_backend/app/requirements.txt
aiofiles==0.7.0
alembic==1.7.4
asgiref==3.4.1
asyncpg==0.24.0
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
coverage==6.0.1
cryptography==3.4.8
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.68.1
greenlet==1.1.1
h11==0.12.0
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
iniconfig==1.1.1
Mako==1.1.5
MarkupSafe==2.0.1
packaging==21.0
passlib==1.7.4
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
pluggy==1.0.0
psycopg2==2.9.1
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==3.0.0
python-dotenv==0.19.0
python-jose==3.3.0
python-magic==0.4.24
python-multipart==0.0.5
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.23
SQLAlchemy-Utils==0.37.8
starlette==0.14.2
tenacity==8.0.1
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.1
urllib3==1.26.6
uvicorn==0.15.0
Werkzeug==2.0.1
zipp==3.6.0
